So first 4 sections
So i did the first 4 sections of the website. i still have some cleaning up to do but this the type of them i was going for. Please text me if you have any ideas. This is not my finished work .. also let me know when you push anything so we do not have any errors.